Add Service Layer for handling bad requests
The server layer will enabel us to validate the client requests
and also perfore the business logic

+++ src/main/java/com/dev/blogpostsapi/data/repository/BlogRepository.java
@@ -3,6 +3,6 @@ package com.dev.blogpostsapi.data.repository;
 import com.dev.blogpostsapi.data.entity.Blog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlogRepository extends JpaRepository<Blog,Long> {
+public interface BlogRepository extends JpaRepository<Blog,Integer> {
     Blog findByBlogTitle(String title);
 }
